# tmpswagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HandySales API - Principal",
    "description": "API REST principal para el sistema HandySales CRM - Gestión de clientes, productos, usuarios y ventas",
    "termsOfService": "https://handysales.com/terms",
    "contact": {
      "name": "HandySales Support",
      "url": "https://handysales.com/support",
      "email": "support@handysales.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0.0"
  },
  "paths": {
...
    "/clientes/{id}": {
      "delete": {
        "tags": [
          "HandySales.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "HandySales.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "HandySales.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/clientes": {
      "get": {
        "tags": [
          "HandySales.Api"
        ],
        "parameters": [
          {
            "name": "ZonaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "CategoriaClienteId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Busqueda",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Activo",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Pagina",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "TamanoPagina",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "HandySales.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/pedidos/cliente/{clienteId}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Pedidos por cliente",
        "description": "Obtiene todos los pedidos de un cliente específico.",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PedidoListaDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/visitas/cliente/{clienteId}": {
      "get": {
        "tags": [
          "Visitas"
        ],
        "summary": "Visitas por cliente",
        "description": "Obtiene el historial de visitas a un cliente específico.",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteVisitaListaDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/sync/pull": {
      "get": {
        "tags": [
          "HandySales.Api"
        ],
        "description": "Pull-only sync to get server changes without pushing local changes.",
        "operationId": "SyncPull",
        "parameters": [
          {
            "name": "lastSyncTimestamp",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "entityTypes",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/sync/status": {
      "get": {
        "tags": [
          "HandySales.Api"
        ],
        "description": "Get sync service status and supported entity types.",
        "operationId": "SyncStatus",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/sync": {
      "post": {
        "tags": [
          "HandySales.Api"
        ],
        "description": "Bidirectional sync for mobile devices. Pushes client changes and pulls server changes.",
        "operationId": "Sync",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/impersonation/end": {
      "post": {
        "tags": [
          "Impersonation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndImpersonationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClienteCreateDto": {
        "required": [
          "correo",
          "direccion",
          "nombre",
          "rfc",
          "telefono"
        ],
        "type": "object",
        "properties": {
          "tenandId": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "rfc": {
            "minLength": 1,
            "type": "string"
          },
          "correo": {
            "minLength": 1,
            "type": "string"
          },
          "telefono": {
            "minLength": 1,
            "type": "string"
          },
          "direccion": {
            "minLength": 1,
            "type": "string"
          },
          "idZona": {
            "type": "integer",
            "format": "int32"
          },
          "categoriaClienteId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ClienteVisitaListaDto": {
        "required": [
          "clienteNombre"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "clienteNombre": {
            "type": "string",
            "nullable": true
          },
          "clienteDireccion": {
            "type": "string",
            "nullable": true
          },
          "fechaProgramada": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaHoraInicio": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaHoraFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "tipoVisita": {
            "$ref": "#/components/schemas/TipoVisita"
          },
          "tipoVisitaNombre": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "resultado": {
            "$ref": "#/components/schemas/ResultadoVisita"
          },
          "resultadoNombre": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "duracionMinutos": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tienePedido": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "EndImpersonationRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "EstadoPedido": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PedidoListaDto": {
        "required": [
          "clienteNombre",
          "numeroPedido",
          "usuarioNombre"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "numeroPedido": {
            "type": "string",
            "nullable": true
          },
          "clienteNombre": {
            "type": "string",
            "nullable": true
          },
          "usuarioNombre": {
            "type": "string",
            "nullable": true
          },
          "fechaPedido": {
            "type": "string",
            "format": "date-time"
          },
          "fechaEntregaEstimada": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoPedido"
          },
          "estadoNombre": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "cantidadProductos": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ResultadoVisita": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SyncChangesDto": {
        "type": "object",
        "properties": {
          "clientes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncClienteDto"
            },
            "nullable": true
          },
          "pedidos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncPedidoDto"
            },
            "nullable": true
          },
          "visitas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncVisitaDto"
            },
            "nullable": true
          },
          "rutas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncRutaDto"
            },
            "nullable": true
          },
          "productos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncProductoDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SyncClienteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "localId": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "rfc": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "idZona": {
            "type": "integer",
            "format": "int32"
          },
          "categoriaClienteId": {
            "type": "integer",
            "format": "int32"
          },
          "latitud": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitud": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "actualizadoEn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "operation": {
            "$ref": "#/components/schemas/SyncOperation"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SyncConflictDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "nullable": true
          },
          "entityId": {
            "type": "integer",
            "format": "int32"
          },
          "field": {
            "type": "string",
            "nullable": true
          },
          "clientValue": {
            "nullable": true
          },
          "serverValue": {
            "nullable": true
          },
          "clientModified": {
            "type": "string",
            "format": "date-time"
          },
          "serverModified": {
            "type": "string",
            "format": "date-time"
          },
          "resolution": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SyncErrorDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "nullable": true
          },
          "entityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "operation": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SyncOperation": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "SyncPedidoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "localId": {
            "type": "string",
            "nullable": true
          },
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "numeroPedido": {
            "type": "string",
            "nullable": true
          },
          "fechaPedido": {
            "type": "string",
            "format": "date-time"
          },
          "fechaEntregaEstimada": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "estado": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "descuento": {
            "type": "number",
            "format": "double"
          },
          "impuestos": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "notas": {
            "type": "string",
            "nullable": true
          },
          "direccionEntrega": {
            "type": "string",
            "nullable": true
          },
          "latitud": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitud": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "listaPrecioId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detalles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncDetallePedidoDto"
            },
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "actualizadoEn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "operation": {
            "$ref": "#/components/schemas/SyncOperation"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SyncRequestDto": {
        "type": "object",
        "properties": {
          "lastSyncTimestamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "entityTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "clientChanges": {
            "$ref": "#/components/schemas/SyncChangesDto"
          }
        },
        "additionalProperties": false
      },
      "SyncResponseDto": {
        "type": "object",
        "properties": {
          "serverTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "serverChanges": {
            "$ref": "#/components/schemas/SyncChangesDto"
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncConflictDto"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncErrorDto"
            },
            "nullable": true
          },
          "summary": {
            "$ref": "#/components/schemas/SyncSummaryDto"
          }
        },
        "additionalProperties": false
      },
...
      "SyncSummaryDto": {
        "type": "object",
        "properties": {
          "clientesPulled": {
            "type": "integer",
            "format": "int32"
          },
          "clientesPushed": {
            "type": "integer",
            "format": "int32"
          },
          "pedidosPulled": {
            "type": "integer",
            "format": "int32"
          },
          "pedidosPushed": {
            "type": "integer",
            "format": "int32"
          },
          "visitasPulled": {
            "type": "integer",
            "format": "int32"
          },
          "visitasPushed": {
            "type": "integer",
            "format": "int32"
          },
          "rutasPulled": {
            "type": "integer",
            "format": "int32"
          },
          "rutasPushed": {
            "type": "integer",
            "format": "int32"
          },
          "productosPulled": {
            "type": "integer",
            "format": "int32"
          },
          "conflictsFound": {
            "type": "integer",
            "format": "int32"
          },
          "errorsFound": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SyncVisitaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "localId": {
            "type": "string",
            "nullable": true
          },
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "fechaProgramada": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaHoraInicio": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaHoraFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "latitudInicio": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitudInicio": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latitudFin": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitudFin": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "estado": {
            "type": "integer",
            "format": "int32"
          },
          "notas": {
            "type": "string",
            "nullable": true
          },
          "pedidoId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "resultado": {
            "type": "string",
            "nullable": true
          },
          "fotos": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "actualizadoEn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "operation": {
            "$ref": "#/components/schemas/SyncOperation"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TipoVisita": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese el token JWT en el formato: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}